Refactor heap logging into dedicated module
